Avoid a "conversion changes signedness: 'long' to 'qulonglong'" case
Revision: https://invent.kde.org/utilities/krusader/-/merge_requests/30

+++ krusader/Archive/krarchandler.cpp
@@ -57,7 +57,7 @@ public:
         qApp->processEvents();
     }
 
-    virtual void subJobStarted(const QString & jobTitle, int count) override {
+    virtual void subJobStarted(const QString & jobTitle, qulonglong count) override {
         krApp->startWaiting(jobTitle, count, true);
     }
 
@@ -289,7 +289,7 @@ bool KrArcHandler::unpack(QString archive, const QString& type, const QString& p
     }
 
     // count the files in the archive
-    long count = arcFileCount(archive, type, password, observer);
+    qulonglong count = arcFileCount(archive, type, password, observer);
     if (count == 0)
         return false;   // not supported
     if (count == 1)
@@ -407,7 +407,7 @@ bool KrArcHandler::unpack(QString archive, const QString& type, const QString& p
     return true; // SUCCESS
 }
 
-bool KrArcHandler::test(const QString& archive, const QString& type, const QString& password, KrArcObserver *observer, long count)
+bool KrArcHandler::test(const QString& archive, const QString& type, const QString& password, KrArcObserver *observer, qulonglong count)
 {
     // choose the right packer for the job
     QStringList packer;
@@ -473,7 +473,7 @@ bool KrArcHandler::test(const QString& archive, const QString& type, const QStri
     return true; // SUCCESS
 }
 
-bool KrArcHandler::pack(QStringList fileNames, QString type, const QString& dest, long count, QMap<QString, QString> extraProps, KrArcObserver *observer)
+bool KrArcHandler::pack(QStringList fileNames, QString type, const QString& dest, qulonglong count, QMap<QString, QString> extraProps, KrArcObserver *observer)
 {
     // set the right packer to do the job
     QStringList packer;
